refactor(agentic-use): drop SDK re-export shims for direct imports
Remove the compatibility shims under tests/agentic-use/runtimes/shared that
re-exported promoted agent_eval SDK symbols, and import those generics directly
from nemo_evaluator_sdk.agent_eval (docker, environment, environment_spec,
gating, verify) at their use sites.

Consolidate the remaining NeMo-Platform-only glue into a single module,
shared/platform.py: the run layout with the platform state_dir, task_image_tag
+ platform DockerEnvironmentProvider, the namespaced AgentPhaseSuccessMetric +
VerifierRewardMetric, agent-log/usage parsing and the shared container env,
attempt construction (live + result.json/ResultDirAttemptSource), the live
VERIFY phase, and the agentic-use task loader. shared/ now holds only
platform.py, config.py, and constants.py.

Update orchestrator/workflow/aut runtimes, the package __init__ re-exports, the
runtime tests, and README/COMPLIANCE docs accordingly. 107 tests pass; ruff
clean.
